+++ CHANGES.md
@@ -5,6 +5,7 @@ Changes in 5.4.2
 ----------------
 - Fix "Strengthen allow edit logic for teachers" when TeacherPrograms (admin) in InputFinalGrades.php
 - SQL Set 50 minutes length for School Periods in rosariosis.sql
+- HTML Short Name input is required in GradeLevels.php
 
 Changes in 5.4.1
 ----------------